style: DTO 패키지 분류 (GS-62)
