+++ src/main/java/com/foo/gosucatcher/domain/item/application/dto/request/sub/SubItemUpdateRequest.java
@@ -1,11 +1,11 @@
-package com.foo.gosucatcher.domain.item.application.dto.request;
+package com.foo.gosucatcher.domain.item.application.dto.request.sub;
+
+import com.foo.gosucatcher.domain.item.domain.MainItem;
+import com.foo.gosucatcher.domain.item.domain.SubItem;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
-
-import com.foo.gosucatcher.domain.item.domain.MainItem;
-import com.foo.gosucatcher.domain.item.domain.SubItem;
 
 public record SubItemUpdateRequest(
 	@NotBlank(message = "하위 서비스명은 필수 입력 입니다.")
