+++ src/main/java/com/foo/gosucatcher/domain/item/application/dto/response/main/MainItemsResponse.java
@@ -1,8 +1,8 @@
-package com.foo.gosucatcher.domain.item.application.dto.response;
-
-import java.util.List;
+package com.foo.gosucatcher.domain.item.application.dto.response.main;
 
 import com.foo.gosucatcher.domain.item.domain.MainItem;
+
+import java.util.List;
 
 public record MainItemsResponse(
 	List<MainItemResponse> mainItemsResponse
